+++ tests/foreman/ui_airgun/test_host.py
@@ -53,6 +53,7 @@ from robottelo.constants import (
 )
 from robottelo.datafactory import gen_string
 from robottelo.decorators import tier2, tier3, skip_if, skip_if_not_set, upgrade
+from robottelo.ui.utils import create_fake_host
 
 
 def _get_set_from_list_of_dict(value):
@@ -130,40 +131,6 @@ def module_host_template(module_org, module_loc):
     host_template.create_missing()
     host_template.name = None
     return host_template
-
-
-def create_fake_host(session, host, interface_id=gen_string('alpha'),
-                     global_parameters=None, host_parameters=None, extra_values=None):
-    if extra_values is None:
-        extra_values = {}
-    os_name = u'{0} {1}'.format(
-        host.operatingsystem.name, host.operatingsystem.major)
-    name = host.name if host.name is not None else gen_string('alpha').lower()
-    values = {
-        'host.name': name,
-        'host.organization': host.organization.name,
-        'host.location': host.location.name,
-        'host.lce': ENVIRONMENT,
-        'host.content_view': DEFAULT_CV,
-        'host.puppet_environment': host.environment.name,
-        'operating_system.architecture': host.architecture.name,
-        'operating_system.operating_system': os_name,
-        'operating_system.media_type': 'All Media',
-        'operating_system.media': host.medium.name,
-        'operating_system.ptable': host.ptable.name,
-        'operating_system.root_password': host.root_pass,
-        'interfaces.interface.interface_type': 'Interface',
-        'interfaces.interface.device_identifier': interface_id,
-        'interfaces.interface.mac': host.mac,
-        'interfaces.interface.domain': host.domain.name,
-        'interfaces.interface.primary': True,
-        'interfaces.interface.interface_additional_data.virtual_nic': False,
-        'parameters.global_params': global_parameters,
-        'parameters.host_params': host_parameters,
-    }
-    values.update(extra_values)
-    session.host.create(values)
-    return '{0}.{1}'.format(name, host.domain.name)
 
 
 @tier3
